+++ age/benches/throughput.rs
@@ -1,4 +1,4 @@
-use age::{Encryptor, SecretKey};
+use age::{x25519, Encryptor};
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion, Throughput};
 use criterion_cycles_per_byte::CyclesPerByte;
 use std::io::{self, Write};
@@ -6,7 +6,7 @@ use std::io::{self, Write};
 const KB: usize = 1024;
 
 fn bench(c: &mut Criterion<CyclesPerByte>) {
-    let recipient = SecretKey::generate().to_public();
+    let recipient = x25519::Identity::generate().to_public();
     let mut group = c.benchmark_group("stream");
 
     for size in &[KB, 2 * KB, 4 * KB, 8 * KB, 16 * KB, 128 * KB] {